pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "doc-mcp"
version = "0.1.0"
description = "Documentation MCP server with Playwright-based authentication and crawling"
readme = "README.md"
requires-python = ">=3.11"
license = "MIT"
classifiers = [
  "License :: OSI Approved :: MIT License",
  "Programming Language :: Python :: 3",
  "Programming Language :: Python :: 3.11",
  "Topic :: Documentation",
  "Topic :: Internet :: WWW/HTTP :: Indexing/Search",
]
dependencies = [
  "playwright>=1.43.0",
  "markdownify>=0.12.0",
  "mcp>=1.0.0",
  "PyYAML>=6.0.1",
  "python-dotenv>=1.0.0",
]

[project.scripts]
docmcp-server = "docmcp.main:main"
docmcp-auth = "docmcp.auth_cli:main"
docmcp-crawl = "docmcp.crawl_cli:main"

[tool.setuptools]
include-package-data = true
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]
include = ["docmcp*"]
